AP-2140 Opponents controller
