+++ app/views/providers/application_merits_task/opponents/show.html.erb
