+++ config/locales/en/providers.yml
@@ -254,6 +254,18 @@ en:
       date_client_told_incidents:
         show:
           h1-heading: Latest incident details
+      opponents:
+        show:
+          h1-heading: Opponent details
+          opponent_name: Opponent's name
+          understands_terms_of_court_order: Does the opponent have the mental capacity to understand the terms of a court order?
+          warning_letter_sent: Has a warning letter been sent to the opponent?
+          police_notified: Have the police been notified?
+          bail_conditions_set: Have bail conditions been set?
+          bail_conditions_set_yes: Give details of the bail conditions, including the date they're likely to end
+          police_notified_no: Tell us why the police have not been notified
+          police_notified_yes: Tell us what action they've taken (if any), including dates and the outcome
+          police_notified_blank: Tell us what action the police have taken
     declarations:
       show:
         h1-heading: Declaration
@@ -337,7 +349,6 @@ en:
           html: Email <a href="mailto:%{team_email}">%{team_email}</a> if you think you do have permission to view this page.
         api_details_user_not_found:
           html: Your login details were not found on the CCMS system.  Please check your credentials and try again.  If the issue persists, please email <a href="mailto:%{team_email}">%{team_email}</a>
-
     legal_aid_applications:
       index:
         heading_1: Your applications
@@ -512,18 +523,7 @@ en:
         list_3: is not employed
         radio_no: No, I do not agree
         radio_yes: Yes, I agree
-    opponents:
-      show:
-        h1-heading: Opponent details
-        opponent_name: Opponent's name
-        understands_terms_of_court_order: Does the opponent have the mental capacity to understand the terms of a court order?
-        warning_letter_sent: Has a warning letter been sent to the opponent?
-        police_notified: Have the police been notified?
-        bail_conditions_set: Have bail conditions been set?
-        bail_conditions_set_yes: Give details of the bail conditions, including the date they're likely to end
-        police_notified_no: Tell us why the police have not been notified
-        police_notified_yes: Tell us what action they've taken (if any), including dates and the outcome
-        police_notified_blank: Tell us what action the police have taken
+
     opponent_names:
       show:
         h1-heading: What is the opponent's name?
