+++ config/routes.rb
@@ -201,7 +201,7 @@ Rails.application.routes.draw do
       resource :capital_income_assessment_result, only: %i[show update]
       resource :identify_types_of_income, only: %i[show update]
       resource :identify_types_of_outgoing, only: %i[show update]
-      resource :opponent, only: %i[show update]
+      resource :opponent, only: %i[show update], controller: 'application_merits_task/opponents'
       resource :opponent_name, only: %i[show update]
       resource :date_client_told_incident, only: %i[show update], controller: 'application_merits_task/date_client_told_incidents'
       resource :proceedings_before_the_court, only: %i[show update]
